Moved runner implementation into it's own file
